requirements.txt
Django==4.1.5
djangorestframework==3.14.0
flake8==3.8.3
psycopg2==2.8.6
Pillow==10.0.0
drf-spectacular==0.26.5